refactor(InMemoryFsClient): extract BasicStats to a separate module
